Project Search/requirements.txt

pymupdf==1.22.5


pillow==9.5.0
pytesseract==0.3.10


sentence-transformers==2.2.2


faiss-cpu==1.7.4


numpy==1.26.0
scikit-learn==1.3.2


streamlit==1.30.0


tqdm==4.68.0
